更改存储

+++ src/main/java/qmstore/shop_cart/controller/ShopCartController.java
@@ -25,7 +25,7 @@ public class ShopCartController {
     @GetMapping("/getShopCartByUser")
     public Response getShopCartByUser(@DataAuth User user, @RequestParam(value = "userId") String userId) {
         // 验证身份
-        if (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN)) {
+        if (user != null && (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN))) {
             return shopCartManager.getShopCartByUser(userId);
         }
         return Response.NOT_LOG_IN();
@@ -40,7 +40,7 @@ public class ShopCartController {
     @GetMapping("/getShopCartByGoods")
     public Response getShopCartByGoods(@DataAuth User user, @RequestParam(value = "goodsId") String goodsId) {
         //仅管理员可用
-        if (user.getUserType().equals(DataType.ADMIN)) {
+        if (user != null && user.getUserType().equals(DataType.ADMIN)) {
             return shopCartManager.getShopCartByGoods(goodsId);
         }
         return Response.NOT_LOG_IN();
@@ -56,7 +56,7 @@ public class ShopCartController {
     @PostMapping("/addShopCart")
     public Response addShopCart(@DataAuth User user, @RequestBody ShopCart shopCart) {
         // 验证身份
-        if (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN)) {
+        if (user != null && (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN))) {
             return shopCartManager.addShopCart(shopCart);
         }
         return Response.NOT_LOG_IN();
@@ -72,7 +72,7 @@ public class ShopCartController {
     @GetMapping("/deleteShopCartByUserAndGoodsId")
     public Response deleteShopCartByUserAndGoodsId(@DataAuth User user, @RequestParam("userId") String userId, @RequestParam("goodsId") String goodsId) {
         // 验证身份
-        if (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN)) {
+        if (user != null && (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN))) {
             return shopCartManager.deleteShopCartByUserAndGoodsId(userId, goodsId);
         }
         return Response.NOT_LOG_IN();
@@ -87,7 +87,7 @@ public class ShopCartController {
     @PostMapping("/updateShopCart")
     public Response updateShopCart(@DataAuth User user, @RequestBody ShopCart shopCart){
         // 验证身份
-        if (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN)) {
+        if (user != null && (user.getUserType().equals(DataType.CUSTOMER) || user.getUserType().equals(DataType.ADMIN))) {
             return shopCartManager.updateShopCart(shopCart);
         }
         return Response.NOT_LOG_IN();
